endee-rag-assistant/pyproject.toml
[project]
name = "endee-rag-research-assistant"
version = "0.1.0"
description = "AI Research Paper Assistant with Endee vector retrieval and RAG"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "langchain>=1.2.12",
    "langchain-google-genai>=4.2.1",
    "langchain-text-splitters>=1.1.1",
    "msgpack>=1.1.2",
    "numpy>=2.4.3",
    "pypdf>=6.8.0",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "sentence-transformers>=5.3.0",
    "streamlit>=1.55.0",
    "uvicorn>=0.41.0",
]
